refactor(test): move test helpers
It feels the test code more belonged to kms_test.go.
This has the benefit of removing some noise from between the tests.
So hopefully the tests are easier to read.
